Cleaned up drag/drop implementation.
