+++ Hexagon/Views/TaskCardView.swift
@@ -31,7 +31,6 @@ struct TaskCardView: View {
         .padding(.horizontal, 16)
         .cardStyle()
         .onTapGesture(perform: onTap)
-//        .draggable(reminder)
         .accessibilityElement(children: .combine)
         .accessibilityLabel(reminder.title ?? "Untitled Task")
         .accessibilityHint("Double-tap for more options")
